Extract library hooks and subcomponents from Library
Library (490 lines) becomes a ~190-line composition. Data loading,
detail polling, catalog management, and admin mutations move to
useLibraryData/useVideoDetail/useCatalogs/useVideoActions; the
portable-catalog panel and per-video detail pane become
library/CatalogPanel and library/VideoDetailPane. Admin and library
page tests pass unchanged.
